+++ src/Chirp.Web/Pages/UserProfile.cshtml.cs
@@ -1,4 +1,6 @@
+using Chirp.Core;
 using Chirp.Infrastructure;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 
@@ -6,12 +8,16 @@ namespace Chirp.Razor.Pages;
 
 public class UserProfileModel : PageModel
 {
-    ICheepRepository _service;
-    private Author Author { get; set; }
+    public ICheepRepository _service;
+    public Author Author { get; set; }
     public string Username { get; set; }
+    public List<CheepDTO> Cheeps { get; set; }
     
-    public UserProfileModel(ICheepRepository service)
+    SignInManager<Author> _signInManager;
+    
+    public UserProfileModel(ICheepRepository service, SignInManager<Author> signInManager)
     {
+        _signInManager = signInManager;
         _service = service;
     }
 
@@ -28,10 +34,28 @@ public class UserProfileModel : PageModel
             return LocalRedirect("/");
         }
         
-        
+        List<CheepDTO> _Cheeps = _service.ReadCheep(0, Username, null);
+        Cheeps = _Cheeps.TakeLast(32).ToList();
         
         return Page();
     }
     
+    public async Task<IActionResult> OnPostDeleteUser(string username)
+    {
+        var info = await _signInManager.GetExternalLoginInfoAsync();
+        if (info != null)
+        {
+            await _signInManager.UserManager.RemoveLoginAsync(_service.GetAuthorByEmail(User.Identity.Name),
+                info.LoginProvider, info.ProviderKey);
+        }
+        _service.DeleteAuthor(username); // Anon user
+        await _signInManager.SignOutAsync(); // Log out user
+        return RedirectToPage("Public"); // Go to main
+    }
     
+    public ActionResult OnPostToggleFollow(string self, string follow)
+    {
+        _service.ToggleFollow(self, follow);
+        return RedirectToPage();
+    }
 }
