+++ src/Chirp.Web/Pages/UserTimeline.cshtml.cs
@@ -33,17 +33,4 @@ public class UserTimelineModel : PostablePage
         
         return Page();
     }
-    
-    public async Task<IActionResult> OnPostDeleteUser(string username)
-    {
-        var info = await _signInManager.GetExternalLoginInfoAsync();
-        if (info != null)
-        {
-            await _signInManager.UserManager.RemoveLoginAsync(_service.GetAuthorByEmail(User.Identity.Name),
-                info.LoginProvider, info.ProviderKey);
-        }
-        _service.DeleteAuthor(username); // Anon user
-        await _signInManager.SignOutAsync(); // Log out user
-        return RedirectToPage("Public"); // Go to main
-    }
 }
